Assert UpdatedAtNanos in neighbor cache tests
Changes the neighbor_cache_test.go tests to always assert UpdatedAtNanos.

Completes the assertion of UpdatedAtNanos in every NUD test, a field that was
historically not checked due to the lack of a deterministic, controllable
clock. This is no longer true with the tcpip.Clock interface. While the tests
have been adjusted to use Clock, asserting by the UpdatedAtNanos was neglected.

Fixes #4663

PiperOrigin-RevId: 359582510